+++ src/config.js
@@ -42,7 +42,8 @@ const fetchAvailableTimeSlots = true;
 
 // A maximum number of days forwards during which a booking can be made.
 // This is limited due to Stripe holding funds up to 90 days from the
-// moment they are charged.
+// moment they are charged. Also note that available time slots can only
+// be fetched for 180 days in the future.
 const dayCountAvailableForBooking = 90;
 
 // To pass environment variables to the client app in the build