+++ src/containers/ListingPage/ListingPage.js
@@ -198,6 +198,7 @@ export class ListingPageComponent extends Component {
       sendEnquiryInProgress,
       sendEnquiryError,
       timeSlots,
+      fetchTimeSlotsError,
       categoriesConfig,
       amenitiesConfig,
     } = this.props;
@@ -479,6 +480,7 @@ export class ListingPageComponent extends Component {
                   handleMobileBookModalClose={handleMobileBookModalClose}
                   onManageDisableScrolling={onManageDisableScrolling}
                   timeSlots={timeSlots}
+                  fetchTimeSlotsError={fetchTimeSlotsError}
                 />
               </div>
             </div>